requirements.txt
django==3.0.4
djangorestframework==3.10.3
dj-database-url==0.5.0